Rename name space to Microsoft.DotNet.ShellShim
