+++ test/Microsoft.DotNet.ShellShimMaker.Tests/Microsoft.DotNet.ShellShim.Tests.csproj
